Create main.cpp
In the united states, land is often measured in square feet. In many  other countries, it is measured in square meters. One acre of land is equivalent of 43,560 square feet. A square meter is equivalent to 10.7639 square feet. Write a program that computes and displays the number of square feet and the number of square meters in 0.5 acre of land. Hint: Because a square meter is larger than a square foot, there will be fewer square meters in 0.5 acre than there are square feet.